requirements.txt
asgiref==3.5.0
astroid==2.11.3
backports.zoneinfo==0.2.1
cffi==1.15.0
cryptography==36.0.2
dill==0.3.4
dj-database-url==0.5.0
Django==4.0.3
django-adminlte3==0.1.6
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-on-heroku==1.1.2
django-rest-knox==4.2.0
djangorestframework==3.13.1
drf-extra-fields==3.4.0
drf-writable-nested==0.6.3
gcloud==0.17.0
googleapis-common-protos==1.56.0
httplib2==0.20.4
isort==5.10.1
jws==0.1.3
lazy-object-proxy==1.7.1
mccabe==0.7.0
mysqlclient==2.1.0
oauth2client==3.0.0
Pillow==9.0.1
platformdirs==2.5.2
protobuf==3.20.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.8
pytz==2021.3
rsa==4.8
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.2.0
whitenoise==6.1.0
wrapt==1.14.0